server/requirements.txt
python3-nmap
netifaces
getmac
ipaddress
mac-vendor-lookup
easysnmp
influxdb
psycopg2
speedtest-cli